0.8alpha.0.28.alpha64.2 	STILL NOT WORK!  SIX IMPOSSIBLE ASSERTIONS BEFORE LUNCH!
	Some fairly random 64 bit cleanups in genesis/runtime, to get
	an SBCL alpha that goes all the way through second genesis and
	a runtime that attempts to laod the core thus created.  Note
	that the core file thus produced can best be described as
	"broken"

+++ src/runtime/cheneygc.c
@@ -527,7 +527,7 @@ scav_fdefn(lispobj *where, lispobj object)
 	== (char *)((unsigned long)(fdefn->raw_addr))) {
         scavenge(where + 1, sizeof(struct fdefn)/sizeof(lispobj) - 1);
         fdefn->raw_addr =
-	    (u32)  ((char *) LOW_WORD(fdefn->fun)) + FUN_RAW_ADDR_OFFSET;
+	    ((char *) LOW_WORD(fdefn->fun)) + FUN_RAW_ADDR_OFFSET;
         return sizeof(struct fdefn) / sizeof(lispobj);
     }
     else
